fix some vulture warnings in production code
