+++ rasa/core/featurizers/single_state_featurizer.py
@@ -378,7 +378,7 @@ class BinarySingleStateFeaturizer(SingleStateFeaturizer):
         super().__init__()
         rasa.shared.utils.io.raise_warning(
             f"'{self.__class__.__name__}' is deprecated and "
-            f"will be removed in the future. "
+            f"will be removed in Rasa Open Source 3.0.0. "
             f"It is recommended to use the '{SingleStateFeaturizer.__name__}' instead.",
             category=DeprecationWarning,
             docs=DOCS_URL_MIGRATION_GUIDE,
@@ -407,7 +407,7 @@ class LabelTokenizerSingleStateFeaturizer(SingleStateFeaturizer):
         # does the same thing if nlu pipeline is configured correctly
         rasa.shared.utils.io.raise_warning(
             f"'{self.__class__.__name__}' is deprecated and "
-            f"will be removed in the future. "
+            f"will be removed in Rasa Open Source 3.0.0. "
             f"It is recommended to use the '{SingleStateFeaturizer.__name__}' instead.",
             category=DeprecationWarning,
             docs=DOCS_URL_MIGRATION_GUIDE,
