+++ rasa/core/lock_store.py
@@ -249,9 +249,6 @@ class RedisLockStore(LockStore):
                 f"Omitting provided non-alphanumeric redis key prefix: '{key_prefix}'. Using default '{self.key_prefix}' instead."
             )
 
-    def _get_key_prefix(self) -> Text:
-        return self.key_prefix
-
     def get_lock(self, conversation_id: Text) -> Optional[TicketLock]:
         serialised_lock = self.red.get(self.key_prefix + conversation_id)
         if serialised_lock:
