+++ rasa/core/utils.py
@@ -1,19 +1,11 @@
-import argparse
 import json
 import logging
 import os
-import re
-import sys
-from asyncio import Future
 from decimal import Decimal
-from hashlib import md5, sha1
 from pathlib import Path
 from typing import (
     Any,
-    Callable,
     Dict,
-    Generator,
-    List,
     Optional,
     Set,
     Text,
@@ -21,12 +13,9 @@ from typing import (
     Union,
 )
 
-import aiohttp
 import numpy as np
 
 import rasa.shared.utils.io
-import rasa.utils.io as io_utils
-from aiohttp import InvalidURL
 from rasa.constants import DEFAULT_SANIC_WORKERS, ENV_SANIC_WORKERS
 from rasa.shared.constants import DEFAULT_ENDPOINTS_PATH
 
@@ -63,18 +52,6 @@ def configure_file_logging(
     logger_obj.addHandler(file_handler)
 
 
-def is_int(value: Any) -> bool:
-    """Checks if a value is an integer.
-
-    The type of the value is not important, it might be an int or a float."""
-
-    # noinspection PyBroadException
-    try:
-        return value == int(value)
-    except Exception:
-        return False
-
-
 def one_hot(hot_idx: int, length: int, dtype: Optional[Text] = None) -> np.ndarray:
     """Create a one-hot array.
 
@@ -94,56 +71,6 @@ def one_hot(hot_idx: int, length: int, dtype: Optional[Text] = None) -> np.ndarr
     r = np.zeros(length, dtype)
     r[hot_idx] = 1
     return r
-
-
-# noinspection PyPep8Naming
-class HashableNDArray:
-    """Hashable wrapper for ndarray objects.
-
-    Instances of ndarray are not hashable, meaning they cannot be added to
-    sets, nor used as keys in dictionaries. This is by design - ndarray
-    objects are mutable, and therefore cannot reliably implement the
-    __hash__() method.
-
-    The hashable class allows a way around this limitation. It implements
-    the required methods for hashable objects in terms of an encapsulated
-    ndarray object. This can be either a copied instance (which is safer)
-    or the original object (which requires the user to be careful enough
-    not to modify it)."""
-
-    def __init__(self, wrapped: np.ndarray, tight: bool = False) -> None:
-        """Creates a new hashable object encapsulating an ndarray.
-
-        wrapped
-            The wrapped ndarray.
-
-        tight
-            Optional. If True, a copy of the input ndaray is created.
-            Defaults to False.
-        """
-
-        self.__tight = tight
-        self.__wrapped = np.array(wrapped) if tight else wrapped
-        self.__hash = int(sha1(wrapped.view()).hexdigest(), 16)  # nosec
-
-    def __eq__(self, other: Any) -> bool:
-        """Performs equality of the underlying array."""
-        return np.all(self.__wrapped == other.__wrapped)
-
-    def __hash__(self) -> int:
-        """Return the hash of the array."""
-        return self.__hash
-
-    def unwrap(self) -> np.ndarray:
-        """Returns the encapsulated ndarray.
-
-        If the wrapper is "tight", a copy of the encapsulated ndarray is
-        returned. Otherwise, the encapsulated ndarray itself is returned."""
-
-        if self.__tight:
-            return np.array(self.__wrapped)
-
-        return self.__wrapped
 
 
 def dump_obj_as_yaml_to_file(
@@ -232,81 +159,10 @@ def is_limit_reached(num_messages: int, limit: Optional[int]) -> bool:
     return limit is not None and num_messages >= limit
 
 
-def read_lines(
-    filename: Union[Path, Text],
-    max_line_limit: Optional[int] = None,
-    line_pattern: Text = ".*",
-) -> Generator[Text, Any, None]:
-    """Read messages from the command line and print bot responses."""
-    line_filter = re.compile(line_pattern)
-
-    with open(filename, "r", encoding=rasa.shared.utils.io.DEFAULT_ENCODING) as f:
-        num_messages = 0
-        for line in f:
-            m = line_filter.match(line)
-            if m is not None:
-                yield m.group(1 if m.lastindex else 0)
-                num_messages += 1
-
-            if is_limit_reached(num_messages, max_line_limit):
-                break
-
-
 def file_as_bytes(path: Text) -> bytes:
     """Read in a file as a byte array."""
     with open(path, "rb") as f:
         return f.read()
-
-
-def convert_bytes_to_string(data: Union[bytes, bytearray, Text]) -> Text:
-    """Convert `data` to string if it is a bytes-like object."""
-
-    if isinstance(data, (bytes, bytearray)):
-        return data.decode(rasa.shared.utils.io.DEFAULT_ENCODING)
-
-    return data
-
-
-def get_file_hash(path: Text) -> Text:
-    """Calculate the md5 hash of a file."""
-    return md5(file_as_bytes(path)).hexdigest()  # nosec
-
-
-async def download_file_from_url(url: Text) -> Text:
-    """Download a story file from a url and persists it into a temp file.
-
-    Args:
-        url: url to download from
-
-    Returns:
-        The file path of the temp file that contains the
-        downloaded content.
-    """
-    from rasa.nlu import utils as nlu_utils
-
-    if not nlu_utils.is_url(url):
-        raise InvalidURL(url)
-
-    async with aiohttp.ClientSession() as session:
-        async with session.get(url, raise_for_status=True) as resp:
-            filename = io_utils.create_temporary_file(await resp.read(), mode="w+b")
-
-    return filename
-
-
-def pad_lists_to_size(
-    list_x: List, list_y: List, padding_value: Optional[Any] = None
-) -> Tuple[List, List]:
-    """Compares list sizes and pads them to equal length."""
-
-    difference = len(list_x) - len(list_y)
-
-    if difference > 0:
-        return list_x, list_y + [padding_value] * difference
-    elif difference < 0:
-        return list_x + [padding_value] * (-difference), list_y
-    else:
-        return list_x, list_y
 
 
 class AvailableEndpoints:
@@ -362,24 +218,6 @@ def read_endpoints_from_path(
         endpoints_path, "endpoints", DEFAULT_ENDPOINTS_PATH, True
     )
     return AvailableEndpoints.read_endpoints(endpoints_config_path)
-
-
-def create_task_error_logger(error_message: Text = "") -> Callable[[Future], None]:
-    """Error logger to be attached to a task.
-
-    This will ensure exceptions are properly logged and won't get lost."""
-
-    def handler(fut: Future) -> None:
-        # noinspection PyBroadException
-        try:
-            fut.result()
-        except Exception:
-            logger.exception(
-                "An exception was raised while running task. "
-                "{}".format(error_message)
-            )
-
-    return handler
 
 
 def replace_floats_with_decimals(obj: Any, round_digits: int = 9) -> Any:
