+++ rasa/shared/core/events.py
@@ -111,18 +111,6 @@ def format_message(
     )
 
 
-def first_key(d: Dict[Text, Any], default_key: Any) -> Any:
-    if len(d) > 1:
-        for k in d.keys():
-            if k != default_key:
-                # we return the first key that is not the default key
-                return k
-    elif len(d) == 1:
-        return list(d.keys())[0]
-    else:
-        return None
-
-
 def split_events(
     events: Iterable["Event"],
     event_type_to_split_on: Type["Event"],
