+++ rasa/shared/core/training_data/story_reader/yaml_story_reader.py
@@ -44,7 +44,6 @@ KEY_ENTITIES = "entities"
 KEY_USER_INTENT = "intent"
 KEY_USER_MESSAGE = "user"
 KEY_SLOT_NAME = "slot_was_set"
-KEY_SLOT_VALUE = "value"
 KEY_ACTIVE_LOOP = "active_loop"
 KEY_ACTION = "action"
 KEY_BOT_END_TO_END_MESSAGE = "bot"
