+++ rasa/shared/nlu/training_data/training_data.py
@@ -465,7 +465,11 @@ class TrainingData:
 
     def sorted_intent_examples(self) -> List[Message]:
         """Sorts the intent examples by the name of the intent and then response"""
-
+        rasa.shared.utils.io.raise_warning(
+            "`sorted_intent_examples` is deprecated and will be removed in Rasa "
+            "3.0.0. ",
+            category=DeprecationWarning,
+        )
         return sorted(
             self.intent_examples,
             key=lambda e: (e.get(INTENT), e.get(INTENT_RESPONSE_KEY)),
