+++ rasa/utils/plotting.py
@@ -21,7 +21,7 @@ def _fix_matplotlib_backend() -> None:
         try:
             # on OSX sometimes the tkinter package is broken and can't be imported.
             # we'll try to import it and if it fails we will use a different backend
-            import tkinter  # skipcq: PYL-W0611
+            import tkinter  # noqa: 401
         except (ImportError, ModuleNotFoundError):
             logger.debug("Setting matplotlib backend to 'agg'")
             matplotlib.use("agg")
@@ -30,7 +30,7 @@ def _fix_matplotlib_backend() -> None:
     elif matplotlib.get_backend() is None:  # pragma: no cover
         try:
             # If the `tkinter` package is available, we can use the `TkAgg` backend
-            import tkinter  # skipcq: PYL-W0611
+            import tkinter  # noqa: 401
 
             logger.debug("Setting matplotlib backend to 'TkAgg'")
             matplotlib.use("TkAgg")
