+++ tests/core/test_utils.py
@@ -19,14 +19,6 @@ class CustomRedisLockStore(RedisLockStore):
     """Test class used to test the behavior of custom lock stores."""
 
 
-def test_is_int():
-    assert utils.is_int(1)
-    assert utils.is_int(1.0)
-    assert not utils.is_int(None)
-    assert not utils.is_int(1.2)
-    assert not utils.is_int("test")
-
-
 def test_one_hot():
     r = utils.one_hot(4, 6)
     assert (r[[0, 1, 2, 3, 5]] == 0).all()
@@ -36,36 +28,6 @@ def test_one_hot():
 def test_one_hot_out_of_range():
     with pytest.raises(ValueError):
         utils.one_hot(4, 3)
-
-
-def test_read_lines():
-    lines = utils.read_lines(
-        "data/test_yaml_stories/stories.yml", max_line_limit=2, line_pattern=r"\.*"
-    )
-
-    lines = list(lines)
-
-    assert len(lines) == 2
-
-
-def test_pad_lists_to_size():
-    list_x = [1, 2, 3]
-    list_y = ["a", "b"]
-    list_z = [None, None, None]
-
-    assert utils.pad_lists_to_size(list_x, list_y) == (list_x, ["a", "b", None])
-    assert utils.pad_lists_to_size(list_y, list_x, "c") == (["a", "b", "c"], list_x)
-    assert utils.pad_lists_to_size(list_z, list_x) == (list_z, list_x)
-
-
-def test_convert_bytes_to_string():
-    # byte string will be decoded
-    byte_string = b"\xcf\x84o\xcf\x81\xce\xbdo\xcf\x82"
-    decoded_string = "τoρνoς"
-    assert utils.convert_bytes_to_string(byte_string) == decoded_string
-
-    # string remains string
-    assert utils.convert_bytes_to_string(decoded_string) == decoded_string
 
 
 @pytest.mark.parametrize(
