+++ tests/nlu/test_utils.py
@@ -36,12 +36,6 @@ def fake_model_dir(empty_model_dir):
     return empty_model_dir  # not empty anymore ;)
 
 
-def test_relative_normpath():
-    test_file = "/my/test/path/file.txt"
-    assert utils.relative_normpath(test_file, "/my/test") == Path("path/file.txt")
-    assert utils.relative_normpath(None, "/my/test") is None
-
-
 def test_list_files_invalid_resource():
     with pytest.raises(ValueError) as execinfo:
         rasa.shared.utils.io.list_files(None)
@@ -66,42 +60,6 @@ def test_list_files_ignores_hidden_files(tmpdir):
 def test_creation_of_existing_dir(tmpdir):
     # makes sure there is no exception
     assert rasa.shared.utils.io.create_directory(tmpdir.strpath) is None
-
-
-def test_empty_is_model_dir(empty_model_dir):
-    assert utils.is_model_dir(empty_model_dir)
-
-
-def test_non_existent_folder_is_no_model_dir():
-    assert not utils.is_model_dir("nonexistent_for_sure_123/")
-
-
-def test_data_folder_is_no_model_dir():
-    assert not utils.is_model_dir("data/")
-
-
-def test_model_folder_is_model_dir(fake_model_dir):
-    assert utils.is_model_dir(fake_model_dir)
-
-
-def test_remove_model_empty(empty_model_dir):
-    assert utils.remove_model(empty_model_dir)
-
-
-def test_remove_model_with_files(fake_model_dir):
-    assert utils.remove_model(fake_model_dir)
-
-
-def test_remove_model_invalid(empty_model_dir):
-    test_file = "something.else"
-    test_content = "Some other stuff"
-    test_file_path = os.path.join(empty_model_dir, test_file)
-    utils.write_to_file(test_file_path, test_content)
-
-    with pytest.raises(RasaException):
-        utils.remove_model(empty_model_dir)
-
-    os.remove(test_file_path)
 
 
 @pytest.mark.parametrize(
