+++ tests/shared/core/training_data/story_reader/test_markdown_story_reader.py
@@ -59,7 +59,7 @@ async def test_persist_and_read_test_story_graph(tmp_path: Path, domain: Domain)
 async def test_persist_and_read_test_story(tmp_path: Path, domain: Domain):
     graph = await training.extract_story_graph("data/test_stories/stories.md", domain)
     out_path = tmp_path / "persisted_story.md"
-    Story(graph.story_steps).dump_to_file(str(out_path))
+    rasa.shared.utils.io.write_text_file(graph.as_story_string(), out_path, append=True)
 
     recovered_trackers = await training.load_data(
         str(out_path),
